[refactor] Created hook for fetchScreening (#54)
