+++ .mailmap
@@ -18,6 +18,10 @@ Cryonato <barisduck@gmail.com> Cryonato <barisduck@gmail.com>
 Markus Wathle <markus.wathle@gmail.com> Markus Wathle <markus.wathle@gmail.com>
 Markus Wathle <markus.wathle@gmail.com> Markus Helgø Wathle <145480355+markuswathle@users.noreply.github.com>
 
+# Helene
+Helene Selvig <helenekselvig@gmail.com> Helene Kolderup Selvig <126758675+HeleneSelvig@users.noreply.github.com>
+Helene Selvig <helenekselvig@gmail.com> Helene Selvig <helenekselvig@gmail.com>
+
 # Ingvild
 Ingvild Kirkaune Sandven <ingviksa@stud.ntnu.no> Ingvild Kirkaune Sandven <ingviksa@stud.ntnu.no>
 Ingvild Kirkaune Sandven <ingviksa@stud.ntnu.no> Ingvild Sandven <ingviksa@stud.ntnu.no>
