+++ frontend/src/api/fetchScreenings.ts
@@ -26,11 +26,14 @@ export async function getScreeningHistory(): Promise<ScreeningDetails[]> {
   return result;
 }
 
-export async function getScreeningByJobPostId(jobPostId: number): Promise<ScreeningDetails[]> {
+export async function getScreeningByJobPostId(
+  jobPostId: number,
+): Promise<ScreeningDetails> {
   const response = await fetch(API_URL + `/api/results/job_posts/${jobPostId}`);
   if (!response.ok) {
     throw new Error(`Response Status: ${response.status}`);
   }
-  const result = await response.json();
+  const result: ScreeningDetails = await response.json();
+
   return result;
 }
