+++ frontend/src/components/PdfPreviewOverlay.tsx
@@ -37,17 +37,14 @@ function PdfPreviewOverlay({
     return <div>Loading...</div>;
   }
 
-  console.log(candidates);
 
   function nextPdfPreview() {
-    console.log(selectedCandidate, isError);
     if (selectedIndex < candidates.length - 1) {
       setSelectedIndex((prev) => prev + 1);
     }
   }
 
   function prevPdfPreview() {
-    console.log(selectedCandidate, isError);
     if (selectedIndex > 0) {
       setSelectedIndex((prev) => prev - 1);
     }
